+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Changed
+ - `AirmoreSession::is_server_running` now uses socket instead of `requests`
+ for faster, accurate and consistent check of Airmore server
+
+### Added
+ - `AirmoreSession::is_authorized` property, which uses the same logic for
+ old `AirmoreSession::is_server_running` property
+
 ## [0.1.0a11] - 2018-10-17
 ### Added
  - "Services" chapter for documentation